requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
python-jose==3.3.0
pydantic==2.5.2
python-dotenv==1.0.0
aioredis==2.0.1
streamlit==1.29.0
httpx==0.25.2
streamlit-authenticator==0.2.3
google-auth-oauthlib==1.0.0
plotly==5.18.0
pandas==2.1.0
streamlit-option-menu==0.3.6
streamlit-extras==0.3.6 